move gradient code to gradient module
